refactoring codes for center stage
:Release Notes:
refactoring codes for center stage

:Detailed Notes:
implement armnn delegate option for cached network
add model param unit test
remove EdgeTPU on default build option
cleanup sources

:Testing Performed:
Local test

:QA Notes:
N/A

:Issues Addressed:
[WRN-16395] integrate refactoring ai framework codes and center stage test app

Reviewed-on: http://gpro.lge.com/c/webos-pro/edge-ai-computer-vision/+/327067
Tested-by: <kijoong.lee@lge.com>

+++ framework/face/FaceParam.cpp
@@ -195,7 +195,7 @@ t_aif_status FaceParam::fromJson(const std::string& param)
         if (modelParam.HasMember("optAspectRatios")) {
             optAspectRatios.clear();
             for (auto& optAspectRatio : modelParam["optAspectRatios"].GetArray()) {
-                optAspectRatios.push_back(optAspectRatio.GetInt());
+                optAspectRatios.push_back(optAspectRatio.GetFloat());
             }
         }
