+++ framework/palm/PalmParam.cpp
@@ -197,7 +197,7 @@ t_aif_status PalmParam::fromJson(const std::string& param)
         if (modelParam.HasMember("optAspectRatios")) {
             optAspectRatios.clear();
             for (auto& optAspectRatio : modelParam["optAspectRatios"].GetArray()) {
-                optAspectRatios.push_back(optAspectRatio.GetInt());
+                optAspectRatios.push_back(optAspectRatio.GetFloat());
             }
         }
